Refactor medical inventory and user avatar request DTOs
- Reintroduced `MedicalInventoryRequest` with full namespace and properties.
- Replaced `UpdateAvatarUrlRequest` with `UserAvatarUrlRequest`.
- Updated import statements in `MedicalInventoryController` and `UsersProfileController` for clarity.
- Ensured `MedicalInventoryService` uses the new DTO namespace.
