+++ SchoolMedicalServer.Abstractions/Dtos/MedicalInventory/MedicalInventoryRequest.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SchoolMedicalServer.Abstractions.Dtos
+namespace SchoolMedicalServer.Abstractions.Dtos.MedicalInventory
 {
     public class MedicalInventoryRequest
     {
